+++ LittleCompany/config/ModConfig.cs
@@ -47,6 +47,8 @@ namespace LittleCompany.Config
 
             public int shrinkRayCost { get; set; }
 
+            public int shrinkRayShotsPerCharge { get; set; }
+
             public float movementSpeedMultiplier { get; set; }
 
             public float jumpHeightMultiplier { get; set; }
@@ -119,6 +121,7 @@ namespace LittleCompany.Config
         public void Setup()
         {
             values.shrinkRayCost                 = Plugin.BepInExConfig().Bind("General", "ShrinkRayCost", 1000, "Store cost of the shrink ray").Value;
+            values.shrinkRayShotsPerCharge       = Plugin.BepInExConfig().Bind("General", "ShrinkRayShotsPerCharge", 7, "Amount of shots per charge for the shrink ray. Set to 0 for unlimited").Value;
             values.deathShrinking                = Plugin.BepInExConfig().Bind("General", "DeathShrinking", false, "If true, a player can be shrunk below 0.2f, resulting in an instant death.").Value;
             values.shrinkRayTargetHighlighting   = Plugin.BepInExConfig().Bind("General", "ShrinkRayTargetHighlighting", ShrinkRayTargetHighlighting.OnHit, "Defines, when a target gets highlighted. Set to OnLoading if you encounter performance issues.").Value;